chore(llm): move provider tests to integration tests 🚚✅
