+++ modules/implementation/tests/timeout.rs
@@ -4,10 +4,10 @@ use std::collections::BTreeMap;
 use std::sync::Arc;
 use tokio::io::AsyncWriteExt;
 
-use super::config::ScriptBackendConfig;
-use super::*;
-use crate::common;
-use crate::scripting;
+use genvm_modules::common;
+use genvm_modules::llm::config::ScriptBackendConfig;
+use genvm_modules::llm::*;
+use genvm_modules::scripting;
 
 #[tokio::test]
 async fn test_timeout() {
@@ -101,7 +101,7 @@ async fn test_timeout() {
         providers: providers.clone(),
         metrics,
     };
-    let sub_ctx = sync::DArc::new(crate::manager::execution_context::LlmSubContext {
+    let sub_ctx = sync::DArc::new(genvm_modules::manager::execution_context::LlmSubContext {
         scripting: scripting_ctx,
         module: llm_ctx,
     });
